+++ world/block.py
@@ -1,4 +1,4 @@
-from world.entity.material import Material
+from world.material.material import Material
 
 
 class Block:
