+++ world/gen/generator.py
@@ -2,7 +2,7 @@ from opensimplex import OpenSimplex
 
 from world.block import Block
 from world.chunk import Chunk
-from world.entity.materials import Materials
+from world.material.materials import Materials
 
 import json
