+++ world/material/materials.py
@@ -3,7 +3,7 @@ from enum import Enum
 import pygame
 
 import assets
-from world.entity.material import Material
+from world.material.material import Material
 
 
 class Materials(Enum):
